backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
motor>=3.6,<4
pydantic-settings>=2.4,<3
pytest>=8.3,<9
httpx>=0.27,<1
PyJWT>=2.10,<3
